+++ packages/seo-plugin/package.json
@@ -47,6 +47,7 @@
 		"@vueuse/core": "^12.5.0",
 		"lodash-es": "^4.17.21",
 		"micromustache": "^8.0.3",
+		"reka-ui": "^2.2.0",
 		"vue-i18n": "^10.0.5"
 	},
 	"devDependencies": {
